pyproject.toml
[project]
name = "sembench-data-import"
version = "1.0.0"
description = "SemBench Movies dataset import tool for PostgreSQL"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.4",
    "tqdm>=4.66.1",
]

[project.scripts]
sembench-import = "importer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["importer.py"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

